+++ cpp/include/copc-lib/las/points.hpp
@@ -114,6 +114,23 @@ class Points
             points_[i]->Z(in[i]);
     }
 
+    std::vector<uint8_t> Classification() const
+    {
+        std::vector<uint8_t> out;
+        out.resize(Size());
+        std::transform(points_.begin(), points_.end(), out.begin(),
+                       [](const std::shared_ptr<Point> &p) { return p->Classification(); });
+        return out;
+    }
+    void Classification(const std::vector<uint8_t> &in)
+    {
+        if (in.size() != Size())
+            throw std::runtime_error("Classification setter array must be same size as Points array!");
+
+        for (unsigned i = 0; i < points_.size(); ++i)
+            points_[i]->Classification(in[i]);
+    }
+
     // Function that return true only if all points are within the box
     bool Within(const Box &box) const
     {